% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sGda}
\alias{sGda}
\title{Prediction function}
\usage{
sGda(data = NULL, data.new = NULL, lam = 0)
}
\arguments{
\item{data}{data matrix with column names being the class labels and row names being the genes.}

\item{data.new}{the new data needs to be predicted.}

\item{lam}{optimal \code{lambda} from cross-validation.}
}
\value{
returns a list object with following items.
\item{pred}{predictions for class labels on the new dataset}
\item{lik}{likelihood of each class on the new dataset}
}
\description{
Prediction function
}

