% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sGdaCV2}
\alias{sGdaCV2}
\title{Cross-validation function}
\usage{
sGdaCV2(data = NULL, cv = 5, lam = 0)
}
\arguments{
\item{data}{data matrix with column names being the class labels and row names being the genes.}

\item{cv}{cross-validation folds.}

\item{lam}{a sequence of \code{lambda}'s.}
}
\value{
returns a list object with following item.
\item{cv.error}{cross-validation errors for each \code{lambda}}
}
\description{
Cross-validation function
}

