% Generated by roxygen2 (4.0.2): do not edit by hand
\name{simpleAGG3}
\alias{simpleAGG3}
\title{Blockwise classifiers}
\usage{
simpleAGG3(data = NULL, data.new = NULL, len = 100, times = 100,
  lam = seq(0, 0.1, length = 10))
}
\arguments{
\item{data}{data matrix with column names being the class labels and row names being the genes.}

\item{data.new}{the new data needs to be predicted.}

\item{len}{block size}

\item{times}{number of blocks}

\item{lam}{a sequence of \code{lambda}'s from cross-validation.}
}
\value{
returns a list object with following items.
\item{cv.error}{cross-validation errors for each block}
\item{pred}{predictions for class labels on the new dataset}
\item{lik}{likelihood of each class on the new dataset}
}
\description{
Blockwise classifiers
}

