% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sortgene}
\alias{sortgene}
\title{Gene sorter}
\usage{
sortgene(data = NULL)
}
\arguments{
\item{data}{data matrix with column names being the class labels and row names being the genes.}
}
\value{
topTable data structure from limma.
}
\description{
Gene sorter
}

